미쳐버린마이페이지-userno 수정

+++ src/main/java/com/venture/suyaho/controller/AdminController.java
@@ -45,7 +45,7 @@ public class AdminController {
                 case "schoolnum":
                     List<User> users = userRepository.findByUserSchoolNumContaining(keyword);
                     if (!users.isEmpty()) {
-                        List<Integer> userListNum = users.stream().map(User::getUserNo).collect(Collectors.toList());
+                        List<Long> userListNum = users.stream().map(User::getUserNo).collect(Collectors.toList());
                         return adminBoardRepository.findByUser_UserNoIn(userListNum);
                     } else {
                         return new ArrayList<>();
