+++ src/main/java/com/venture/suyaho/repository/AdminBoardRepository.java
@@ -20,7 +20,7 @@ public interface AdminBoardRepository extends JpaRepository<AdminBoard, Long> {
 
     // 사용자 번호 리스트로 게시글 검색
     @Query("SELECT ab FROM AdminBoard ab JOIN FETCH ab.user WHERE ab.user.userNo IN :userListNum")
-    List<AdminBoard> findByUser_UserNoIn(@Param("userListNum") List<Integer> userListNum);
+    List<AdminBoard> findByUser_UserNoIn(@Param("userListNum") List<Long> userListNum);
 
     List<AdminBoard> findByUser(User user);
